requirement.txt
 streamlit
 google-generativeai
 python-dotenv
 psycopg2-binary
 psycopg2
